exp check energy check weight check

+++ src/main/java/com/catholic/meowlife/domain/entity/CatEntity.java
@@ -13,6 +13,7 @@ public class CatEntity {
     private String catName;
     private String catBreeds;
     private int energy;
-    private int weight;
+    private double weight;
     private int exp;
 }
+
